+++ CHANGELOG.md
@@ -31,4 +31,4 @@
 
 ### Features
 
-* initial release of Argus MCP Server ([16e6910](https://github.com/RaphaEnterprises-AI/argus-mcp-server/commit/16e6910358a14e468899f5a78d04b7809b52262b))
+* initial release of Skopaq MCP Server ([16e6910](https://github.com/RaphaEnterprises-AI/argus-mcp-server/commit/16e6910358a14e468899f5a78d04b7809b52262b))